# provider: Reduce schema memory usage by 18 MB