--- a/internal/service/ec2/tags.go
+++ b/internal/service/ec2/tags.go
@@ -100,8 +100,8 @@ func tagsFromTagDescriptions(tds []awstypes.TagDescription) []awstypes.Tag {
 }
 
 func tagsSchemaConflictsWith(conflictsWith []string) *schema.Schema {
-	v := tftags.TagsSchema()
+	v := *tftags.TagsSchema()
 	v.ConflictsWith = conflictsWith
 
-	return v
+	return &v
 }