--- a/internal/service/ec2/tags.go
+++ b/internal/service/ec2/tags.go
@@ -100,7 +100,7 @@ func tagsFromTagDescriptions(tds []awstypes.TagDescription) []awstypes.Tag {
 }
 
 func tagsSchemaConflictsWith(conflictsWith []string) *schema.Schema {
-	v := *tftags.TagsSchema()
+	v := *tftags.TagsSchema() // nosemgrep:ci.semgrep.aws.prefer-pointer-conversion-assignment
 	v.ConflictsWith = conflictsWith
 
 	return &v