### Description

Reduces schema memory usage by approximately 18 MB and eliminates approximately 117,600 allocations

1. Shares Plugin SDK tag schemas across resources and data sources. Saves approximately 1 MB and eliminates approximately 3,100 allocations.

Before: BenchmarkSDKProviderInitialization-10 40 29284118 ns/op 70426029 B/op 448276 allocs/op
After: BenchmarkSDKProviderInitialization-10 40 28709038 ns/op 69434669 B/op 445177 allocs/op

2. Shares repeated WAFv2 schemas. Saves approximately 17 MB and eliminates approximately 114,500 allocations.

Before: BenchmarkSDKProviderInitialization-10 40 28709038 ns/op 69434669 B/op 445177 allocs/op
After: BenchmarkSDKProviderInitialization-10 50 22838247 ns/op 52270798 B/op 330692 allocs/op

### Relations
<!---
If your pull request fully resolves and should automatically close the linked issue, use Closes. Otherwise, use Relates.

For Example:

Relates #0000
or 
Closes #0000
--->

Relates #31722